+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Adding support for plural selection clauses in both `xan select -e` & `xan map` e.g. `xan map 'full_name.split(" ") as (first_name, last_name)`.
 * Adding `xan search -P/--add-pattern`.
 * Adding `xan groupby -M/--along-matrix`.
+* Adding `xan groupby -T/--total`.
 
 *Fixes*